+++ docs/changelog.rst
@@ -199,6 +199,12 @@ you use a decent Wayland compositor.
 Detailed list of changes
 -------------------------------------
 
+0.49.2 [future]
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- Fix an out-of-bounds memory access when drawing a multi-line text sized
+  character with the cursor below the bottom margin of the scroll region
+
 0.49.1 [2026-09-24]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~